requirements.txt
fastapi
typing 
pydantic
requests
rasterio
uvicorn
python-multipart
sklearn.cluster
matplotlib
base64